Fully working TLOP and added move function to chassis

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/odometry/position/Position.java
@@ -3,7 +3,7 @@ package org.firstinspires.ftc.teamcode.odometry.position;
 public class Position {
     private static double xPosition;
     private static double yPosition;
-    private static double orientation;
+    private static int orientation;
 
     public Position(double xPosition, double yPosition, int orientation) {
         Position.xPosition = xPosition;
@@ -17,11 +17,11 @@ public class Position {
     public static double getYPosition(){
         return yPosition;
     }
-    public static double getOrientation(){
+    public static int getOrientation(){
         return orientation;
     }
 
     public static void setXPosition(double xposition){Position.xPosition = xposition;}
     public static void setYPosition(double yposition){Position.yPosition = yposition;}
-    public static void setOrientation(double orientation){Position.orientation = orientation;}
+    public static void setOrientation(int orientation){Position.orientation = orientation;}
 }
